Adding Readme file.
Satan McNice.

A vertical slice of a noir 2D point-n-click adventure made in Unity as a learning experience during the fall of 2016. It was created by me (Oscar Mårtensson) and Marcus Johansson.

The game is about Satan McNice - a hard-boiled cop - who arrives at Old McDonald's farm to investigate a reported crime.